+++ mRemoteNG.Specs/packages.config
@@ -1,18 +1,18 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="BoDi" version="1.4.1" targetFramework="net46" />
+  <package id="BoDi" version="1.4.1" targetFramework="net472" />
   <package id="Cucumber.Messages" version="6.0.2" targetFramework="net472" />
-  <package id="Gherkin" version="6.0.0" targetFramework="net46" />
+  <package id="Gherkin" version="6.0.0" targetFramework="net472" />
   <package id="Google.Protobuf" version="3.12.1" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net472" />
-  <package id="NUnit" version="3.12.0" targetFramework="net46" />
+  <package id="NUnit" version="3.12.0" targetFramework="net472" />
   <package id="NUnit.Console" version="3.11.1" targetFramework="net472" />
   <package id="NUnit.ConsoleRunner" version="3.11.1" targetFramework="net472" />
-  <package id="NUnit.Extension.NUnitProjectLoader" version="3.6.0" targetFramework="net46" />
+  <package id="NUnit.Extension.NUnitProjectLoader" version="3.6.0" targetFramework="net472" />
   <package id="NUnit.Extension.NUnitV2Driver" version="3.8.0" targetFramework="net472" />
-  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.6.0" targetFramework="net46" />
+  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.6.0" targetFramework="net472" />
   <package id="NUnit.Extension.TeamCityEventListener" version="1.0.7" targetFramework="net472" />
-  <package id="NUnit.Extension.VSProjectLoader" version="3.8.0" targetFramework="net46" />
+  <package id="NUnit.Extension.VSProjectLoader" version="3.8.0" targetFramework="net472" />
   <package id="NUnit.Runners" version="3.11.1" targetFramework="net472" />
   <package id="NUnit3TestAdapter" version="3.16.1" targetFramework="net472" developmentDependency="true" />
   <package id="SpecFlow" version="3.1.97" targetFramework="net472" />
@@ -32,6 +32,6 @@
   <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net472" />
   <package id="System.Security.Cryptography.X509Certificates" version="4.3.2" targetFramework="net472" />
   <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net46" requireReinstallation="true" />
-  <package id="Utf8Json" version="1.3.7" targetFramework="net46" requireReinstallation="true" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net472" />
+  <package id="Utf8Json" version="1.3.7" targetFramework="net472" />
 </packages>
