+++ mRemoteNGTests/packages.config
@@ -2,20 +2,20 @@
 <packages>
   <package id="BouncyCastle" version="1.8.6.1" targetFramework="net472" />
   <package id="Castle.Core" version="4.4.1" targetFramework="net472" />
-  <package id="DockPanelSuite" version="3.0.6" targetFramework="net46" />
-  <package id="log4net" version="2.0.8" targetFramework="net46" />
+  <package id="DockPanelSuite" version="3.0.6" targetFramework="net472" />
+  <package id="log4net" version="2.0.8" targetFramework="net472" />
   <package id="NSubstitute" version="4.2.1" targetFramework="net472" />
-  <package id="NUnit" version="3.12.0" targetFramework="net46" />
+  <package id="NUnit" version="3.12.0" targetFramework="net472" />
   <package id="NUnit.Console" version="3.11.1" targetFramework="net472" />
   <package id="NUnit.ConsoleRunner" version="3.11.1" targetFramework="net472" />
-  <package id="NUnit.Extension.NUnitProjectLoader" version="3.6.0" targetFramework="net46" />
+  <package id="NUnit.Extension.NUnitProjectLoader" version="3.6.0" targetFramework="net472" />
   <package id="NUnit.Extension.NUnitV2Driver" version="3.8.0" targetFramework="net472" />
-  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.6.0" targetFramework="net46" />
+  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.6.0" targetFramework="net472" />
   <package id="NUnit.Extension.TeamCityEventListener" version="1.0.7" targetFramework="net472" />
-  <package id="NUnit.Extension.VSProjectLoader" version="3.8.0" targetFramework="net46" />
+  <package id="NUnit.Extension.VSProjectLoader" version="3.8.0" targetFramework="net472" />
   <package id="NUnit3TestAdapter" version="3.16.1" targetFramework="net472" developmentDependency="true" />
-  <package id="ObjectListView.Official" version="2.9.1" targetFramework="net46" />
-  <package id="OpenCover" version="4.7.922" targetFramework="net46" />
+  <package id="ObjectListView.Official" version="2.9.1" targetFramework="net472" />
+  <package id="OpenCover" version="4.7.922" targetFramework="net472" />
   <package id="ReportGenerator" version="4.5.8" targetFramework="net472" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net472" />
   <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
