+++ mRemoteV1/packages.config
@@ -3,22 +3,21 @@
   <package id="BouncyCastle" version="1.8.6.1" targetFramework="net472" />
   <package id="ConsoleControl" version="1.3.0" targetFramework="net472" />
   <package id="ConsoleControlAPI" version="1.3.0" targetFramework="net472" />
-  <package id="DockPanelSuite" version="3.0.6" targetFramework="net46" />
-  <package id="DockPanelSuite.ThemeVS2003" version="3.0.6" targetFramework="net46" />
-  <package id="DockPanelSuite.ThemeVS2012" version="3.0.6" targetFramework="net46" />
-  <package id="DockPanelSuite.ThemeVS2013" version="3.0.6" targetFramework="net46" />
-  <package id="DockPanelSuite.ThemeVS2015" version="3.0.6" targetFramework="net46" />
+  <package id="DockPanelSuite" version="3.0.6" targetFramework="net472" />
+  <package id="DockPanelSuite.ThemeVS2003" version="3.0.6" targetFramework="net472" />
+  <package id="DockPanelSuite.ThemeVS2012" version="3.0.6" targetFramework="net472" />
+  <package id="DockPanelSuite.ThemeVS2013" version="3.0.6" targetFramework="net472" />
+  <package id="DockPanelSuite.ThemeVS2015" version="3.0.6" targetFramework="net472" />
   <package id="Geckofx60.32" version="60.0.44" targetFramework="net472" />
   <package id="Google.Protobuf" version="3.12.1" targetFramework="net472" />
   <package id="K4os.Compression.LZ4" version="1.1.11" targetFramework="net472" />
   <package id="K4os.Compression.LZ4.Streams" version="1.1.11" targetFramework="net472" />
   <package id="K4os.Hash.xxHash" version="1.0.6" targetFramework="net472" />
-  <package id="log4net" version="2.0.8" targetFramework="net46" />
-  <package id="Markdig" version="0.20.0" targetFramework="net472" />
+  <package id="log4net" version="2.0.8" targetFramework="net472" />
   <package id="MySql.Data" version="8.0.20" targetFramework="net472" />
   <package id="NUnit3TestAdapter" version="3.16.1" targetFramework="net472" developmentDependency="true" />
-  <package id="ObjectListView.Official" version="2.9.1" targetFramework="net46" />
-  <package id="SSH.NET" version="2016.1.0" targetFramework="net46" />
+  <package id="ObjectListView.Official" version="2.9.1" targetFramework="net472" />
+  <package id="SSH.NET" version="2016.1.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
   <package id="System.Memory" version="4.5.4" targetFramework="net472" />
   <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
